# Check for duplicated series on a scrape