@roidelapluie 2023-10-04
Since official client libraries will not allow that, is it worth the extra performance penalty?