@jesusvazquez 2023-10-31
This PR was reviewed during the Prometheus Bug Scrub event. 

We'd like to run Prombench on this pull request to check for possible performance regressions. Unfortunately Prombench is currently not working so we have to fix it first. We'll come back to this PR once it is fixed.