@bboreham 2023-11-12
Seems considerably more expensive in CPU.
<img width="430" alt="image" src="https://github.com/prometheus/prometheus/assets/8125524/2636d9ff-ec02-4651-8db6-2835fb745453">

and memory
<img width="430" alt="image" src="https://github.com/prometheus/prometheus/assets/8125524/f32f1c7b-35ee-4ff7-813a-98e24495ab1d">

Prometheus is already tracking which series sent samples, for staleness.  Maybe that can be used for the same purpose?